Reset noise retriggering counter upon noise retrigger _off_ command
Prior to this change the counter was never reset and its state
depended on accumulated duration and parameters of previous noise
retriggering commands making it hard to predict the number of ticks
before the next trigger when setting retrigger to _on_. Resetting
the counter upon retrigger _off_ command leaves open the possibility
of updating retrigger parameters without affecting the counter.
Achieving a known next trigger delay is done by having a retrigger
_off_ command followed immediately by a retrigger _on_ command
(which get applied in the same tick).