Reset glissando counter upon glissando _off_ command
Prior to this change the counter was never reset and its the state
depended on accumulated duration and parameters of previous glissando
commands making it hard to predict the number of ticks before the next
pitch update when setting glissando _on_. Resetting upon glissando
_off_ leaves open the possibility of updating glissando parameters
without affecting the counter. Achieving a known glissando delay
is done by having a glissando _off_ command followed immediately by a
glissando _on_ command (which get applied in the same tick).