Implement new command format and command parser to reduce progmem usage
WARNING: commands format has changed and it is not backwards compatible.

The number parameter bytes for some commands (LFO and SLIDE) has increased
by one. This is more than offset by the reduction in code size. The new
command format now supports appling LFO effects to the square wave duty
cycle (unimplmented) and to specify the LFO waveform (only the original
triangular LFO is implemented now).

A set of constants and macros was added to aid manual creation of scores
and the simple song+sfx example has been updated to show how to use them.